+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.3.1-SNAPSHOT
 
+* Document version retention now uses a watermark approach by which old
+  versions of a document can be deleted in bulk once the retention threshold is
+  reached.
+
 * The default consistent hashing selector now uses a FNV hashing algorithm,
   replacing the use of Murmur3. Better performance, no collisions for millions
   of keys, better key distribution across nodes. Utils.computeHash uses FNV